refactor: Update `Maze._break_walls_r` and it's tests
This commit will update the `Maze._break_walls_r` method as follows:
> Refer the method docstring for overall explanation because the
following points just talk about the partial functionality that changed.

Before:
-> Find the neighbour of Current Cell and randomly choose one of them to
go to if they haven't been visited.
-> In case all of the neighbours of Current Cell have been visited, mark
any one of those neighbours as Current Cell and repeat.

This approach tries to randomly find a Cell on Maze to start building a
new maze path in case the existing path hits a roadblock, i.e., all the
neighbours of current Cell have been visited.

After:
-> Find the neighbour of Current Cell and randomly choose one of them to
go to if they haven't been visited.
-> In case all of the neighbours of Current Cell have been visited,
backtrack to previous Cell marking it as Current Cell and repeat to
traverse a new maze path.

This approach follows the DFS (Depth-First Search) algorithm to find the
maze path in case it hits a roadblock.

This will not affect the UI of the app.